requirements.txt
groq==0.18.0
PyPDF2==3.0.1
reportlab==4.0.8
python-dotenv   
requests
pinecone-client
sentence-transformers
pandas==2.0.3
openpyxl==3.1.2
flask==2.3.3
flask-cors==4.0.0
gunicorn